minitap/ai_agent_menu/0_intro_adk/pyproject.toml
[project]
name = "gcp-getting-started-lab-jp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
    "google-cloud-aiplatform>=1.121.0",
    "google-genai>=1.45.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]
